# perf(serde_v8): fast path for large strings