```
Before:
test de_sob_str_6b        ... bench:          48 ns/iter (+/- 2)
test de_sob_str_64b       ... bench:         105 ns/iter (+/- 4)
test de_sob_str_512b      ... bench:         327 ns/iter (+/- 19)
test de_sob_str_1kb       ... bench:         460 ns/iter (+/- 5)
test de_sob_str_4kb       ... bench:       1,571 ns/iter (+/- 27)

After:
test de_sob_str_6b        ... bench:          49 ns/iter (+/- 3)
test de_sob_str_64b       ... bench:          66 ns/iter (+/- 9)
test de_sob_str_512b      ... bench:         119 ns/iter (+/- 3)
test de_sob_str_1kb       ... bench:          84 ns/iter (+/- 3)
test de_sob_str_4kb       ... bench:         181 ns/iter (+/- 8)
```